+++ client/src/lib/forms/invalid-callbacks.js
@@ -139,6 +139,16 @@ function invalidEncryptionKeyRing(stateData) {
 }
 
 /**
+ * invalid encryption key endpoint
+ * @param {Object} stateData
+ * @param {Object} stateData.endpoint
+ * @return {boolean} true if invalid
+ */
+function invalidEncryptionKeyEndpoint(stateData) {
+  return stateData.endpoint === "public-and-private";
+}
+
+/**
  * validate sshKey
  * @param {string} str
  * @returns {boolean} true if it is a valid sshKey
@@ -680,6 +690,7 @@ module.exports = {
   invalidName,
   invalidNewResourceName,
   invalidEncryptionKeyRing,
+  invalidEncryptionKeyEndpoint,
   invalidSshPublicKey,
   invalidTagList,
   invalidCrnList,